rename KafkaClient to LegacyKafkaClient and extract interface
